+++ MafiaAPI/Models/Boss.cs
@@ -17,7 +17,7 @@ namespace MafiaAPI.Models
 
         public string LastName { get; set; }
         public string FirstName { get; set; }
-        public int Money { get; set; }
+        public long Money { get; set; }
         public DateTime? LastSeen { get; set; }
 
         public virtual Player Player { get; set; }
